Move concat into hooks module
